+++ crates/ironrdp-testsuite-core/tests/server/mod.rs
@@ -1,1 +1,2 @@
+mod acceptor;
 mod fast_path;
